refactor: split unpublished-commit provenance out of fix_cycle
fix_cycle.py had grown to 1532 lines, over the 1500-line first-party
limit enforced by
tests/unit/test_core_decomposition_maintainability.py::
test_first_party_code_files_stay_under_line_limit.

Move the self-contained unpublished-commit provenance bookkeeping —
_git_push_result_with_terminal_head_provenance_unavailable,
_git_push_result_with_local_terminal_head and
_enrich_failed_fix_cycle_result — into the sibling module
fix_cycle_preserved_commit.py and re-export
_enrich_failed_fix_cycle_result from fix_cycle so existing callers and
tests keep their import path. Code is moved verbatim; behaviour is
unchanged.

fix_cycle.py is now 1423 lines, leaving headroom for the remaining
review fixes on this PR.
